backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
torch==2.0.1
transformers==4.33.0
peft==0.4.0
bitsandbytes==0.41.1
accelerate==0.23.0